3 Assignment Pinecone/requirements.txt
langgraph
langchain
langchain-core
langchain-google-genai
pinecone-client
google-cloud-aiplatform
python-dotenv
pydantic
